refactor(groups): migrate content-sharing routes to ts-rest contract
Move the group content-sharing endpoints (share/unshare, list content,
permissions, remove, vorlagen) from the legacy raw groupContent.ts handlers
onto the groupsContract, and repoint every frontend caller (useGroupSharing
plus the ~8 POST /share sites across texts/recherche/notebooks/boards/agents/
image-studio/group-info) to the typed contracts client. Behaviour and response
shapes are unchanged.

The polymorphic GET /content response is modelled as a fixed bucket envelope
with loose per-item records (the frontend already narrows buckets via its own
casts). groupContent.ts is now an empty router kept for composition. Avatar
(multipart) stays on the legacy router.
