+++ apps/api/routes/auth/groups/groupContent.ts
@@ -1,953 +1,14 @@
 /**
- * Group content sharing routes
- * Handles content sharing, permissions, and group content retrieval
+ * Group content-sharing routes — MIGRATED.
+ *
+ * All content-sharing endpoints (share/unshare, list content, permissions,
+ * remove, vorlagen) were moved to the ts-rest contract router
+ * (`groupsContractRouter.ts`). This module now exports only an empty Express
+ * router, kept mounted for composition compatibility (`groups/index.ts`).
  */
 
-import fs from 'fs';
-import path from 'path';
-
-import express, { type Router, type Response } from 'express';
-
-import { NotebookQdrantHelper } from '../../../database/services/NotebookQdrantHelper.js';
-import authMiddlewareModule from '../../../middleware/authMiddleware.js';
-import { validateBody, type TypedRequest } from '../../../middleware/validateBody.js';
-import { NextcloudShareManager } from '../../../utils/integrations/nextcloud/index.js';
-import { createLogger } from '../../../utils/logger.js';
-import { fromParam, type GroupId } from '../../../utils/types/branded.js';
-import {
-  groupContentShareSchema,
-  groupContentUnshareSchema,
-  groupContentPermissionsSchema,
-  groupContentDeleteSchema,
-  type AuthRequest,
-  type GroupContentShareBody,
-  type GroupContentUnshareBody,
-  type GroupContentPermissionsBody,
-  type GroupContentDeleteBody,
-} from '../types.js';
-
-import { getPostgresAndCheckMembership } from './groupCore.js';
-
-const log = createLogger('groupContent');
-const { requireAuth: ensureAuthenticated } = authMiddlewareModule;
-
-const notebookHelper = new NotebookQdrantHelper();
+import express, { type Router } from 'express';
 
 const router: Router = express.Router();
-
-interface ShareRecord {
-  content_type: string;
-  content_id: string;
-  shared_at: string;
-  permissions: string | Record<string, unknown>;
-  shared_by_user_id: string;
-  first_name: string | null;
-  display_name: string | null;
-}
-
-interface ContentItem {
-  id: string;
-  [key: string]: unknown;
-}
-
-interface SystemTemplate {
-  id: string;
-  title: string;
-  description: string;
-  template_type: string;
-  thumbnail_url: string;
-  preview_image?: string;
-  external_url: string;
-  tags: string[];
-  categories: string[];
-}
-
-let systemTemplates: SystemTemplate[] = [];
-try {
-  const apiRoot = process.cwd();
-  const systemTemplatesPath = path.resolve(apiRoot, 'config/templates/system-templates.json');
-  const data = fs.readFileSync(systemTemplatesPath, 'utf-8');
-  const parsed = JSON.parse(data) as {
-    templates?: Array<SystemTemplate & { preview_image?: string }>;
-  };
-  systemTemplates = (parsed.templates ?? []).map(
-    (t: SystemTemplate & { preview_image?: string }) => ({
-      ...t,
-      thumbnail_url: t.preview_image
-        ? `/auth/template-previews/${t.preview_image}`
-        : t.thumbnail_url,
-    })
-  );
-} catch {
-  log.warn('[Group Content] Could not load system templates for vorlagen matching');
-}
-
-// ============================================================================
-// Group Vorlagen (tag-based template matching)
-// ============================================================================
-
-router.get(
-  '/groups/:groupId/vorlagen',
-  ensureAuthenticated,
-  async (req: AuthRequest<{ groupId: string }>, res: Response): Promise<void> => {
-    try {
-      const groupId = fromParam<GroupId>(req.params.groupId);
-      const userId = req.user!.id;
-
-      const { postgres } = await getPostgresAndCheckMembership(groupId, userId, false);
-
-      // Read group settings for templateTags
-      const group = await postgres.queryOne(
-        'SELECT settings FROM groups WHERE id = $1',
-        [groupId],
-        { table: 'groups' }
-      );
-
-      const settings =
-        typeof group?.settings === 'string'
-          ? (JSON.parse(group.settings) as { templateTags?: string[] })
-          : ((group?.settings as { templateTags?: string[] } | null) ?? {});
-      const templateTags: string[] = settings.templateTags ?? [];
-
-      if (templateTags.length === 0) {
-        res.json({ success: true, vorlagen: [], tags: [] });
-        return;
-      }
-
-      // Query published user templates where tags overlap with group tags
-      // Using ?| operator: "does the JSONB array contain any of these values?"
-      const dbTemplates = await postgres.query(
-        `SELECT id, title, description, template_type, thumbnail_url, external_url,
-                tags, categories, metadata, created_at
-         FROM user_templates
-         WHERE is_private = false
-           AND status = 'published'
-           AND type = 'template'
-           AND tags ?| $1::text[]
-         ORDER BY created_at DESC`,
-        [templateTags],
-        { table: 'user_templates' }
-      );
-
-      // Filter system templates that match any of the group tags
-      const lowerTags = templateTags.map((t: string) => t.toLowerCase());
-      const matchingSystemTemplates = systemTemplates
-        .filter((t) => {
-          const tTags = (t.tags || []).map((tag: string) => tag.toLowerCase());
-          const tCategories = (t.categories || []).map((c: string) => c.toLowerCase());
-          const tType = (t.template_type || '').toLowerCase();
-          return lowerTags.some(
-            (groupTag) =>
-              tTags.includes(groupTag) || tCategories.includes(groupTag) || tType === groupTag
-          );
-        })
-        .map((t) => ({
-          ...t,
-          is_system: true,
-        }));
-
-      // Combine and deduplicate by id
-      const seenIds = new Set<string>();
-      const allVorlagen: Array<{
-        id: string;
-        title: string;
-        description: string;
-        template_type: string;
-        thumbnail_url: string;
-        external_url: string;
-        is_system: boolean;
-        tags?: string[];
-        categories?: string[];
-        created_at?: unknown;
-      }> = [];
-
-      for (const t of [
-        ...((dbTemplates || []) as Array<Record<string, unknown> & SystemTemplate>),
-        ...matchingSystemTemplates,
-      ]) {
-        if (!seenIds.has(t.id)) {
-          seenIds.add(t.id);
-          allVorlagen.push({
-            id: t.id,
-            title: t.title,
-            description: t.description,
-            template_type: t.template_type,
-            thumbnail_url: t.thumbnail_url,
-            external_url: t.external_url,
-            tags: t.tags || [],
-            categories: ((t as Record<string, unknown>).categories as string[]) || [],
-            is_system: !!t.is_system,
-            created_at: 'created_at' in t ? t.created_at : undefined,
-          });
-        }
-      }
-
-      res.json({
-        success: true,
-        vorlagen: allVorlagen,
-        tags: templateTags,
-      });
-    } catch (error) {
-      const err = error as Error;
-      log.error('[Group Vorlagen] Error:', err);
-      res.status(500).json({
-        success: false,
-        message: err.message || 'Fehler beim Laden der Vorlagen.',
-      });
-    }
-  }
-);
-
-const tableNameMap: Record<string, string> = {
-  database: 'user_templates',
-  template: 'user_templates',
-  user_templates: 'user_templates',
-  instructions: 'user_instructions',
-  user_instructions: 'user_instructions',
-  canvas_template: 'collaborative_documents',
-};
-
-// ============================================================================
-// Share Content Endpoints
-// ============================================================================
-
-// Share content to a group
-router.post(
-  '/groups/:groupId/share',
-  ensureAuthenticated,
-  validateBody(groupContentShareSchema),
-  async (
-    req: TypedRequest<GroupContentShareBody, { groupId: string }>,
-    res: Response
-  ): Promise<void> => {
-    try {
-      const groupId = fromParam<GroupId>(req.params.groupId);
-      const userId = req.user!.id;
-      const { contentType, contentId, permissions } = req.body;
-
-      const { postgres } = await getPostgresAndCheckMembership(groupId, userId, false);
-
-      // Nextcloud share links don't live in their own SQL table — they're entries
-      // in the owner's profiles.nextcloud_share_links JSONB. Verify ownership via
-      // the share manager, which throws if the link isn't in the user's array.
-      if (contentType === 'nextcloud_share_link') {
-        try {
-          await NextcloudShareManager.getShareLinkById(userId, contentId);
-        } catch {
-          res.status(404).json({
-            success: false,
-            message: 'Wolke-Verbindung nicht gefunden.',
-          });
-          return;
-        }
-      }
-
-      // User notebooks live in Qdrant, not Postgres — verify ownership there.
-      if (contentType === 'notebook_collections') {
-        const collection = await notebookHelper.getNotebookCollection(contentId);
-        if (!collection) {
-          res.status(404).json({
-            success: false,
-            message: 'Inhalt nicht gefunden.',
-          });
-          return;
-        }
-        if (collection.user_id !== userId) {
-          res.status(403).json({
-            success: false,
-            message: 'Du bist nicht Besitzer*in dieses Inhalts.',
-          });
-          return;
-        }
-      }
-
-      // System content (notebooks/agents) is globally available — skip ownership check
-      if (
-        contentType !== 'system_notebooks' &&
-        contentType !== 'system_agents' &&
-        contentType !== 'nextcloud_share_link' &&
-        contentType !== 'notebook_collections'
-      ) {
-        const tableName = tableNameMap[contentType] || contentType;
-        const ownerColumn =
-          contentType === 'collaborative_documents' || contentType === 'canvas_template'
-            ? 'created_by'
-            : 'user_id';
-
-        // Build ownership query based on content type
-        let ownershipSQL = `SELECT ${ownerColumn} FROM ${tableName} WHERE id = $1`;
-        const ownershipParams: Array<string> = [contentId];
-
-        // For user_templates table (templates), also filter by type = 'template'
-        if (tableName === 'user_templates') {
-          ownershipSQL += ` AND type = $2`;
-          ownershipParams.push('template');
-        }
-
-        // For collaborative_documents, also filter out deleted
-        if (contentType === 'collaborative_documents') {
-          ownershipSQL += ` AND is_deleted = false`;
-        }
-
-        // For canvas_template (also lives in collaborative_documents), only allow non-deleted canvases
-        if (contentType === 'canvas_template') {
-          ownershipSQL += ` AND is_deleted = false AND document_subtype = 'canvas'`;
-        }
-
-        const contentOwnership = await postgres.queryOne<{ [key: string]: string }>(
-          ownershipSQL,
-          ownershipParams,
-          {
-            table: tableName,
-          }
-        );
-
-        if (!contentOwnership) {
-          log.error(
-            '[User Groups /groups/:groupId/share POST] Content ownership verification failed:',
-            {
-              contentType,
-              contentId,
-              userId,
-            }
-          );
-          res.status(404).json({
-            success: false,
-            message: 'Inhalt nicht gefunden.',
-          });
-          return;
-        }
-
-        if (contentOwnership[ownerColumn] !== userId) {
-          res.status(403).json({
-            success: false,
-            message: 'Du bist nicht Besitzer*in dieses Inhalts.',
-          });
-          return;
-        }
-      }
-
-      const existingShare = await postgres.queryOne<{ id: string }>(
-        'SELECT id FROM group_content_shares WHERE content_type = $1 AND content_id = $2 AND group_id = $3',
-        [contentType, contentId, groupId],
-        { table: 'group_content_shares' }
-      );
-
-      if (existingShare) {
-        res.status(400).json({
-          success: false,
-          message: 'Inhalt ist bereits mit dieser Gruppe geteilt.',
-        });
-        return;
-      }
-
-      const sharePermissions = permissions ?? {
-        read: true,
-        write: false,
-        collaborative: false,
-      };
-
-      log.debug('[User Groups /share] Inserting share record:', {
-        contentType,
-        contentId,
-        groupId,
-        userId,
-        permissions: sharePermissions,
-      });
-
-      await postgres.exec(
-        'INSERT INTO group_content_shares (content_type, content_id, group_id, shared_by_user_id, permissions) VALUES ($1, $2, $3, $4, $5)',
-        [contentType, contentId, groupId, userId, JSON.stringify(sharePermissions)]
-      );
-
-      log.debug('[User Groups /share] Share record inserted successfully');
-
-      const CONTENT_LABELS: Record<string, string> = {
-        documents: 'ein Dokument',
-        custom_generators: 'einen Grünerator',
-        notebook_collections: 'ein Notizbuch',
-        user_documents: 'einen Text',
-        collaborative_documents: 'ein Dokument',
-        database: 'einen Datenbank-Eintrag',
-        system_notebooks: 'ein Notizbuch',
-        system_agents: 'einen Agenten',
-        canvas_template: 'eine Sharepic-Vorlage',
-        nextcloud_share_link: 'eine Wolke-Verbindung',
-      };
-      import('../../../services/notifications/index.js')
-        .then(({ notifyGroupMembers }) => {
-          const groupInfo = postgres.queryOne('SELECT name FROM groups WHERE id = $1', [groupId], {
-            table: 'groups',
-          });
-          return groupInfo.then((g) =>
-            notifyGroupMembers({
-              groupId,
-              excludeUserId: userId,
-              type: 'group_content_shared',
-              title: 'Neuer Inhalt',
-              body: `${req.user?.display_name || 'Jemand'} hat ${CONTENT_LABELS[contentType] || 'etwas'} in „${g?.name || 'deiner Gruppe'}" geteilt`,
-              actionUrl: `/gruppen/${groupId}`,
-              metadata: { contentType, contentId },
-            })
-          );
-        })
-        .catch(() => {});
-
-      res.json({
-        success: true,
-        message: 'Inhalt erfolgreich mit der Gruppe geteilt.',
-      });
-    } catch (error) {
-      const err = error as Error;
-      log.error('[User Groups /groups/:groupId/share POST] Error:', err);
-      res.status(500).json({
-        success: false,
-        message: err.message || 'Fehler beim Teilen des Inhalts.',
-      });
-    }
-  }
-);
-
-// Unshare content from a group
-router.delete(
-  '/groups/:groupId/share',
-  ensureAuthenticated,
-  validateBody(groupContentUnshareSchema),
-  async (
-    req: TypedRequest<GroupContentUnshareBody, { groupId: string }>,
-    res: Response
-  ): Promise<void> => {
-    try {
-      const groupId = fromParam<GroupId>(req.params.groupId);
-      const userId = req.user!.id;
-      const { contentType, contentId } = req.body;
-
-      const { postgres } = await getPostgresAndCheckMembership(groupId, userId, false);
-
-      // Verify the share exists and user owns it or has permission to unshare
-      const shareRecord = await postgres.queryOne<{ shared_by_user_id: string }>(
-        'SELECT shared_by_user_id FROM group_content_shares WHERE content_type = $1 AND content_id = $2 AND group_id = $3',
-        [contentType, contentId, groupId],
-        { table: 'group_content_shares' }
-      );
-
-      if (!shareRecord) {
-        res.status(404).json({
-          success: false,
-          message: 'Geteilter Inhalt nicht gefunden.',
-        });
-        return;
-      }
-
-      // Only the user who shared the content can unshare it (or group admins in future)
-      if (shareRecord.shared_by_user_id !== userId) {
-        res.status(403).json({
-          success: false,
-          message: 'Du kannst nur Inhalte aufheben, die du selbst geteilt hast.',
-        });
-        return;
-      }
-
-      // Remove from junction table
-      const result = await postgres.exec(
-        'DELETE FROM group_content_shares WHERE content_type = $1 AND content_id = $2 AND group_id = $3',
-        [contentType, contentId, groupId]
-      );
-
-      if (result.changes === 0) {
-        throw new Error('Share record not found or already deleted');
-      }
-
-      res.json({
-        success: true,
-        message: 'Inhalt wurde erfolgreich aus der Gruppe entfernt.',
-      });
-    } catch (error) {
-      const err = error as Error;
-      log.error('[User Groups /groups/:groupId/share DELETE] Error:', err);
-      res.status(500).json({
-        success: false,
-        message: err.message || 'Fehler beim Entfernen des Inhalts aus der Gruppe.',
-      });
-    }
-  }
-);
-
-// ============================================================================
-// Get Group Content
-// ============================================================================
-
-// Get all content shared with a group
-router.get(
-  '/groups/:groupId/content',
-  ensureAuthenticated,
-  async (req: AuthRequest<{ groupId: string }>, res: Response): Promise<void> => {
-    try {
-      const groupId = fromParam<GroupId>(req.params.groupId);
-      const userId = req.user!.id;
-
-      const { postgres } = await getPostgresAndCheckMembership(groupId, userId, false);
-
-      // Fetch shared content with user profile information
-      const sharedContent: ShareRecord[] =
-        ((await postgres.query(
-          `
-      SELECT
-        gcs.content_type,
-        gcs.content_id,
-        gcs.shared_at,
-        gcs.permissions,
-        gcs.shared_by_user_id,
-        p.first_name,
-        p.display_name
-      FROM group_content_shares gcs
-      LEFT JOIN profiles p ON p.id = gcs.shared_by_user_id
-      WHERE gcs.group_id = $1
-      ORDER BY gcs.shared_at DESC
-    `,
-          [groupId],
-          { table: 'group_content_shares' }
-        )) as ShareRecord[]) || [];
-
-      log.debug('[User Groups /content] Fetched shared content:', {
-        groupId,
-        totalShares: sharedContent.length,
-        contentTypes: sharedContent.map((s: ShareRecord) => s.content_type),
-      });
-
-      // Group shared content by type for easier processing
-      const contentByType: Record<string, ShareRecord[]> = {
-        documents: [],
-        custom_generators: [],
-        notebook_collections: [],
-        user_documents: [],
-        database: [],
-        collaborative_documents: [],
-        system_notebooks: [],
-        system_agents: [],
-        canvas_template: [],
-      };
-
-      sharedContent.forEach((share: ShareRecord) => {
-        if (contentByType[share.content_type]) {
-          contentByType[share.content_type].push(share);
-        }
-      });
-
-      // Fetch actual content details for all types in parallel
-      type ContentResult = {
-        type: string;
-        result: { data: Array<Record<string, unknown>> };
-        shares: ShareRecord[];
-      };
-      const fetchPromises: Promise<ContentResult | null>[] = [];
-
-      if (contentByType.documents.length > 0) {
-        const ids = contentByType.documents.map((s: ShareRecord) => s.content_id);
-        fetchPromises.push(
-          postgres
-            .query(
-              'SELECT id, title, filename, file_size, status, created_at, updated_at, user_id FROM documents WHERE id = ANY($1)',
-              [ids],
-              { table: 'documents' }
-            )
-            .then((data) => ({
-              type: 'documents',
-              result: { data: data || [] },
-              shares: contentByType.documents,
-            }))
-        );
-      }
-
-      if (contentByType.custom_generators.length > 0) {
-        const ids = contentByType.custom_generators.map((s: ShareRecord) => s.content_id);
-        fetchPromises.push(
-          postgres
-            .query(
-              'SELECT id, name, title, description, created_at, updated_at, user_id FROM custom_generators WHERE id = ANY($1)',
-              [ids],
-              { table: 'custom_generators' }
-            )
-            .then((data) => ({
-              type: 'custom_generators',
-              result: { data: data || [] },
-              shares: contentByType.custom_generators,
-            }))
-        );
-      }
-
-      if (contentByType.notebook_collections.length > 0) {
-        const ids = contentByType.notebook_collections.map((s: ShareRecord) => s.content_id);
-        // Notebooks live in Qdrant, not Postgres — hydrate details from there.
-        fetchPromises.push(
-          notebookHelper.getNotebookCollectionsByIds(ids).then((collections) => ({
-            type: 'notebook_collections',
-            result: {
-              data: collections.map((c) => ({
-                id: c.id,
-                name: c.name,
-                description: c.description,
-                created_at: c.created_at,
-                updated_at: c.updated_at,
-                user_id: c.user_id,
-              })),
-            },
-            shares: contentByType.notebook_collections,
-          }))
-        );
-      }
-
-      // System Notebooks (no DB lookup needed — frontend resolves display from config)
-      if (contentByType.system_notebooks.length > 0) {
-        const systemNotebooksData = contentByType.system_notebooks.map((s: ShareRecord) => ({
-          id: s.content_id,
-          system: true,
-        }));
-        fetchPromises.push(
-          Promise.resolve({
-            type: 'system_notebooks',
-            result: { data: systemNotebooksData },
-            shares: contentByType.system_notebooks,
-          })
-        );
-      }
-
-      // System Agents (no DB lookup needed — frontend resolves display from SYSTEM_AGENTS)
-      if (contentByType.system_agents.length > 0) {
-        const systemAgentsData = contentByType.system_agents.map((s: ShareRecord) => ({
-          id: s.content_id,
-          system: true,
-        }));
-        fetchPromises.push(
-          Promise.resolve({
-            type: 'system_agents',
-            result: { data: systemAgentsData },
-            shares: contentByType.system_agents,
-          })
-        );
-      }
-
-      if (contentByType.user_documents.length > 0) {
-        const ids = contentByType.user_documents.map((s: ShareRecord) => s.content_id);
-        fetchPromises.push(
-          postgres
-            .query(
-              'SELECT id, title, document_type, content, created_at, updated_at, user_id FROM user_documents WHERE id = ANY($1)',
-              [ids],
-              { table: 'user_documents' }
-            )
-            .then((rawData) => {
-              const textsData = ((rawData || []) as Array<ContentItem & { content?: string }>).map(
-                (item) => {
-                  let plainText = item.content || '';
-                  let prev = '';
-                  while (prev !== plainText) {
-                    prev = plainText;
-                    plainText = plainText.replace(/<[^>]*>/g, '');
-                  }
-                  plainText = plainText.trim();
-                  const wordCount = plainText
-                    .split(/\s+/)
-                    .filter((word: string) => word.length > 0).length;
-                  return { ...item, word_count: wordCount, character_count: plainText.length };
-                }
-              );
-              return {
-                type: 'user_documents',
-                result: { data: textsData },
-                shares: contentByType.user_documents,
-              };
-            })
-        );
-      }
-
-      if (contentByType.database.length > 0) {
-        const ids = contentByType.database.map((s: ShareRecord) => s.content_id);
-        fetchPromises.push(
-          postgres
-            .query(
-              "SELECT id, title, description, external_url, thumbnail_url, metadata, created_at, updated_at, user_id FROM user_templates WHERE id = ANY($1) AND type = 'template'",
-              [ids],
-              { table: 'user_templates' }
-            )
-            .then((data) => ({
-              type: 'database',
-              result: { data: data || [] },
-              shares: contentByType.database,
-            }))
-        );
-      }
-
-      if (contentByType.collaborative_documents.length > 0) {
-        const ids = contentByType.collaborative_documents.map((s: ShareRecord) => s.content_id);
-        fetchPromises.push(
-          postgres
-            .query(
-              'SELECT id, title, document_subtype, created_by, created_at, updated_at FROM collaborative_documents WHERE id = ANY($1::uuid[]) AND is_deleted = false',
-              [ids],
-              { table: 'collaborative_documents' }
-            )
-            .then((data) => ({
-              type: 'collaborative_documents',
-              result: { data: data || [] },
-              shares: contentByType.collaborative_documents,
-            }))
-        );
-      }
-
-      if (contentByType.canvas_template.length > 0) {
-        const ids = contentByType.canvas_template.map((s: ShareRecord) => s.content_id);
-        fetchPromises.push(
-          postgres
-            .query(
-              `SELECT cd.id, cd.title, cd.created_by,
-                      cd.created_at, cd.updated_at,
-                      cdoc.template_type, cdoc.thumbnail_url, cdoc.format
-               FROM collaborative_documents cd
-               INNER JOIN canvas_documents cdoc ON cdoc.document_id = cd.id
-               WHERE cd.id = ANY($1::uuid[])
-                 AND cd.is_deleted = false
-                 AND cd.document_subtype = 'canvas'`,
-              [ids],
-              { table: 'collaborative_documents' }
-            )
-            .then((data) => ({
-              type: 'canvas_template',
-              result: { data: data || [] },
-              shares: contentByType.canvas_template,
-            }))
-        );
-      }
-
-      const contentResults = (await Promise.all(fetchPromises)).filter(Boolean) as ContentResult[];
-
-      // Process and format results
-      const groupContent: Record<string, unknown[]> = {
-        documents: [],
-        generators: [],
-        notebooks: [],
-        texts: [],
-        templates: [],
-        collaborative_documents: [],
-        system_notebooks: [],
-        system_agents: [],
-        canvas_templates: [],
-      };
-
-      contentResults.forEach(({ type, result, shares }) => {
-        const items = (result.data || []).map((item) => {
-          // Find the corresponding share info
-          const shareInfo = shares.find((s: ShareRecord) => s.content_id === item.id);
-
-          const parsedPermissions: Record<string, unknown> =
-            typeof shareInfo?.permissions === 'string'
-              ? (JSON.parse(shareInfo.permissions) as Record<string, unknown>)
-              : ((shareInfo?.permissions as Record<string, unknown> | null) ?? {});
-
-          const parsedMetadata: Record<string, unknown> =
-            type === 'database' && item.metadata != null
-              ? typeof item.metadata === 'string'
-                ? (JSON.parse(item.metadata) as Record<string, unknown>)
-                : (item.metadata as Record<string, unknown>)
-              : {};
-
-          return {
-            ...item,
-            contentType: type,
-            shared_at: shareInfo?.shared_at,
-            group_permissions: parsedPermissions,
-            shared_by_name: shareInfo?.display_name || shareInfo?.first_name || 'Unknown User',
-            // Add template-specific fields for database
-            ...(type === 'database' && {
-              template_type: (parsedMetadata.template_type as string) || 'template',
-              external_url: item.external_url,
-            }),
-          };
-        });
-
-        // Map to the correct groupContent key
-        const keyMap: Record<string, string> = {
-          documents: 'documents',
-          custom_generators: 'generators',
-          notebook_collections: 'notebooks',
-          user_documents: 'texts',
-          database: 'templates',
-          collaborative_documents: 'collaborative_documents',
-          system_notebooks: 'system_notebooks',
-          system_agents: 'system_agents',
-          canvas_template: 'canvas_templates',
-        };
-
-        groupContent[keyMap[type]] = items;
-      });
-
-      log.debug('[User Groups /content] Final response:', {
-        templatesCount: groupContent.templates.length,
-        documentsCount: groupContent.documents.length,
-        textsCount: groupContent.texts.length,
-      });
-
-      res.json({
-        success: true,
-        content: groupContent,
-      });
-    } catch (error) {
-      const err = error as Error;
-      log.error('[User Groups /groups/:groupId/content GET] Error:', err);
-      res.status(500).json({
-        success: false,
-        message: err.message || 'Fehler beim Laden der Gruppeninhalte.',
-      });
-    }
-  }
-);
-
-// ============================================================================
-// Content Permissions
-// ============================================================================
-
-// Update content permissions
-router.put(
-  '/groups/:groupId/content/:contentId/permissions',
-  ensureAuthenticated,
-  validateBody(groupContentPermissionsSchema),
-  async (
-    req: AuthRequest<{ groupId: string; contentId: string }> &
-      TypedRequest<GroupContentPermissionsBody>,
-    res: Response
-  ): Promise<void> => {
-    try {
-      const groupId = fromParam<GroupId>(req.params.groupId);
-      const contentId = req.params.contentId;
-      const userId = req.user!.id;
-      const { contentType, permissions } = req.body as {
-        contentType: string;
-        permissions: string[];
-      };
-
-      const { postgres, membership } = await getPostgresAndCheckMembership(groupId, userId, false);
-
-      // Check if content is shared with the group and get share info
-      const shareRecord = await postgres.queryOne<{ shared_by_user_id: string }>(
-        'SELECT shared_by_user_id FROM group_content_shares WHERE content_type = $1 AND content_id = $2 AND group_id = $3',
-        [contentType, contentId, groupId],
-        { table: 'group_content_shares' }
-      );
-
-      if (!shareRecord) {
-        res.status(404).json({
-          success: false,
-          message: 'Inhalt ist nicht mit dieser Gruppe geteilt.',
-        });
-        return;
-      }
-
-      // Check if user has permission to modify permissions (admin or content sharer)
-      const isAdmin = membership.role === 'admin';
-      const isSharer = shareRecord.shared_by_user_id === userId;
-
-      if (!isAdmin && !isSharer) {
-        res.status(403).json({
-          success: false,
-          message: 'Keine Berechtigung zum Ändern der Berechtigungen.',
-        });
-        return;
-      }
-
-      // Update permissions in the junction table
-      const result = await postgres.exec(
-        'UPDATE group_content_shares SET permissions = $1 WHERE content_type = $2 AND content_id = $3 AND group_id = $4',
-        [JSON.stringify(permissions), contentType, contentId, groupId]
-      );
-
-      if (result.changes === 0) {
-        throw new Error('Share record not found or no changes made');
-      }
-
-      res.json({
-        success: true,
-        message: 'Berechtigungen erfolgreich aktualisiert.',
-      });
-    } catch (error) {
-      const err = error as Error;
-      log.error('[User Groups /groups/:groupId/content/:contentId/permissions PUT] Error:', err);
-      res.status(500).json({
-        success: false,
-        message: err.message || 'Fehler beim Aktualisieren der Berechtigungen.',
-      });
-    }
-  }
-);
-
-// Remove content from group (unshare)
-router.delete(
-  '/groups/:groupId/content/:contentId',
-  ensureAuthenticated,
-  validateBody(groupContentDeleteSchema),
-  async (
-    req: TypedRequest<GroupContentDeleteBody, { groupId: string; contentId: string }>,
-    res: Response
-  ): Promise<void> => {
-    try {
-      const groupId = fromParam<GroupId>(req.params.groupId);
-      const contentId = req.params.contentId;
-      const userId = req.user!.id;
-      const { contentType } = req.body;
-
-      const { postgres, membership } = await getPostgresAndCheckMembership(groupId, userId, false);
-
-      // Check if user is admin
-      const isAdmin = membership.role === 'admin';
-
-      // For now, only admins can unshare content (can be extended later)
-      if (!isAdmin) {
-        res.status(403).json({
-          success: false,
-          message: 'Nur Gruppenadministratoren können geteilte Inhalte entfernen.',
-        });
-        return;
-      }
-
-      // Verify the share exists in the junction table
-      const shareRecord = await postgres.queryOne<{ shared_by_user_id: string }>(
-        'SELECT shared_by_user_id FROM group_content_shares WHERE content_type = $1 AND content_id = $2 AND group_id = $3',
-        [contentType, contentId, groupId],
-        { table: 'group_content_shares' }
-      );
-
-      if (!shareRecord) {
-        log.error('[User Groups /groups/:groupId/content/:contentId DELETE] Share check error');
-        res.status(404).json({
-          success: false,
-          message: 'Geteilter Inhalt nicht gefunden.',
-        });
-        return;
-      }
-
-      // Remove from junction table
-      const result = await postgres.exec(
-        'DELETE FROM group_content_shares WHERE content_type = $1 AND content_id = $2 AND group_id = $3',
-        [contentType, contentId, groupId]
-      );
-
-      if (result.changes === 0) {
-        log.error('[User Groups /groups/:groupId/content/:contentId DELETE] Unshare error');
-        throw new Error('Share record not found or already deleted');
-      }
-
-      res.json({
-        success: true,
-        message: 'Inhalt erfolgreich aus der Gruppe entfernt.',
-      });
-    } catch (error) {
-      const err = error as Error;
-      log.error('[User Groups /groups/:groupId/content/:contentId DELETE] Error:', err);
-      res.status(500).json({
-        success: false,
-        message: err.message || 'Fehler beim Entfernen des geteilten Inhalts.',
-      });
-    }
-  }
-);
 
 export default router;
